move context provider to own module
