+++ compass-geocoder-mobile/build.gradle.kts
@@ -16,8 +16,8 @@ kotlin {
         }
 
         androidMain.dependencies {
+            implementation(projects.compassToolsAndroid)
             implementation(libs.kotlinx.coroutines.android)
-            implementation(libs.androidx.startup)
         }
     }
 }
