+++ compass-geocoder-mobile/src/androidMain/AndroidManifest.xml
@@ -1,24 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    >
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
     <uses-permission android:name="android.permission.INTERNET" />
-
-    <application>
-        <provider
-            android:name="androidx.startup.InitializationProvider"
-            android:authorities="${applicationId}.androidx-startup"
-            android:exported="false"
-            tools:node="merge"
-            >
-            <meta-data
-                android:name="dev.jordond.compass.geocoder.mobile.internal.context.ContextProviderInitializer"
-                android:value="androidx.startup"
-                />
-        </provider>
-
-    </application>
 
 </manifest>
