+++ compass-geocoder-mobile/src/androidMain/kotlin/dev/jordond/compass/geocoder/mobile/MobilePlatformGeocoder.android.kt
@@ -7,10 +7,10 @@ import dev.jordond.compass.Coordinates
 import dev.jordond.compass.Place
 import dev.jordond.compass.exception.NotSupportedException
 import dev.jordond.compass.geocoder.mobile.internal.asyncOperation
-import dev.jordond.compass.geocoder.mobile.internal.context.ContextProvider
 import dev.jordond.compass.geocoder.mobile.internal.syncOperation
 import dev.jordond.compass.geocoder.mobile.internal.toCoordinates
 import dev.jordond.compass.geocoder.mobile.internal.toPlaces
+import dev.jordond.compass.tools.ContextProvider
 
 internal class AndroidPlatformGeocoder(
     private val context: Context,
