+++ compass-geolocation-mobile/build.gradle.kts
@@ -15,6 +15,7 @@ kotlin {
         }
 
         androidMain.dependencies {
+            implementation(projects.compassToolsAndroid)
             api(libs.play.services.location)
         }
     }
