+++ compass-geolocation-mobile/src/commonMain/kotlin/dev/jordond/compass/geolocation/mobile/MobileLocator.kt
@@ -1,0 +1,7 @@
+package dev.jordond.compass.geolocation.mobile
+
+import dev.jordond.compass.geolocation.Locator
+
+public interface MobileLocator : Locator
+
+internal expect fun createLocator(): MobileLocator
