--- compass-geolocation-mobile/src/commonMain/kotlin/dev/jordond/compass/geolocation/mobile/MobilePlatformLocator.kt
@@ -1,1 +1,0 @@
-public interface MobilePlatformLocator
