+++ compass-geolocation-mobile/src/iosMain/kotlin/dev/jordond/compass/geolocation/mobile/MobileLocator.ios.kt
@@ -1,0 +1,5 @@
+package dev.jordond.compass.geolocation.mobile
+
+internal actual fun createLocator(): MobileLocator {
+    TODO("Not yet implemented")
+}
