+++ compass-tools-android/src/androidMain/kotlin/dev/jordond/compass/tools/ContextProvider.kt
@@ -1,4 +1,4 @@
-package dev.jordond.compass.geocoder.mobile.internal.context
+package dev.jordond.compass.tools
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -6,14 +6,14 @@ import android.content.Context
 /**
  * Class for providing the application context.
  */
-internal class ContextProvider(internal val context: Context) {
+public class ContextProvider(public val context: Context) {
 
-    companion object {
+    public companion object {
 
         @SuppressLint("StaticFieldLeak")
         private var instance: ContextProvider? = null
 
-        fun create(context: Context): ContextProvider {
+        public fun create(context: Context): ContextProvider {
             if (instance == null) {
                 instance = ContextProvider(context)
             }
@@ -21,7 +21,7 @@ internal class ContextProvider(internal val context: Context) {
             return instance!!
         }
 
-        fun getInstance(): ContextProvider = instance
+        public fun getInstance(): ContextProvider = instance
             ?: throw IllegalStateException("ContextProvider has not been initialized")
     }
 }
